fixed misbehaviour
Added a semaphore to prevent misbehaviour. Everytime the mouse
leaves the slider, the slider will switch to the next slide. So
wiggling the cursor on and off the slider, the slider will rotate
the slides very fast. This has been fixed with a semaphore,
which will alllow only one timeout at the same time.